Route service modified, past event option added, new review function, mappers added and new achivement service

+++ src/main/java/cat/cultura/backend/controller/UserController.java
@@ -1,14 +1,14 @@
 package cat.cultura.backend.controller;
 
-import cat.cultura.backend.dtos.EventDto;
-import cat.cultura.backend.dtos.LoggedUserDto;
-import cat.cultura.backend.dtos.TrophyDto;
-import cat.cultura.backend.dtos.UserDto;
+import cat.cultura.backend.dtos.*;
 import cat.cultura.backend.entity.*;
-import cat.cultura.backend.exceptions.UserNotFoundException;
+import cat.cultura.backend.mappers.ReviewMapper;
+import cat.cultura.backend.mappers.RouteMapper;
+import cat.cultura.backend.mappers.TrophyMapper;
+import cat.cultura.backend.service.*;
+import cat.cultura.backend.service.AttendanceService;
 import cat.cultura.backend.interceptors.CurrentUser;
 import cat.cultura.backend.mappers.EventMapper;
-import cat.cultura.backend.service.*;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -39,10 +39,25 @@ public class UserController {
     private RequestService requestService;
 
     @Autowired
+    private ReviewService reviewService;
+
+    @Autowired
     private EventMapper eventMapper;
 
     @Autowired
+    private ReviewMapper reviewMapper;
+
+    @Autowired
+    private RouteMapper routeMapper;
+
+    @Autowired
+    private TrophyMapper trophyMapper;
+
+    @Autowired
     private ModelMapper modelMapper;
+
+    @Autowired
+    private RouteService routeService;
 
 
     @GetMapping("/login")
@@ -103,15 +118,20 @@ public class UserController {
 
     @PutMapping("/users/{id}/attendance/{eventId}")
     public ResponseEntity<String> addAttendance(@PathVariable Long id, @PathVariable Long eventId) {
-        try {
-            attendanceService.addAttendance(id,eventId);
-        } catch (AssertionError as) {
-            ResponseEntity.status(HttpStatus.BAD_REQUEST).body(as.getMessage());
+        if(!isCurrentUser(id)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        } else {
+            try {
+                attendanceService.addAttendance(id, eventId);
+            } catch (AssertionError as) {
+                ResponseEntity.status(HttpStatus.BAD_REQUEST).body(as.getMessage());
+            }
+            userTrophyService.firstAttendance(id);
+            return ResponseEntity.status(HttpStatus.CREATED).body("Event attendance added");
         }
-        return ResponseEntity.status(HttpStatus.CREATED).body("Event attendance added");
     }
 
-    @GetMapping("/user/{id}/attendance")
+    @GetMapping("/users/{id}/attendance")
     public ResponseEntity<List<EventDto>> getAttendance(@PathVariable Long id) {
         List<Event> events = userService.getAttendanceEventsById(id);
         return ResponseEntity.status(HttpStatus.OK).body(events.stream().map(eventMapper::convertEventToDto).toList());
@@ -119,24 +139,33 @@ public class UserController {
 
     @DeleteMapping("/users/{id}/assistance/{eventId}")
     public ResponseEntity<String> removeAttendance(@PathVariable Long id, @PathVariable Long eventId) {
-        try {
-            attendanceService.removeAttendance(id, eventId);
-        } catch (AssertionError as) {
-            ResponseEntity.status(HttpStatus.BAD_REQUEST).body(as.getMessage());
+        if(!isCurrentUser(id)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        } else {
+            try {
+                attendanceService.removeAttendance(id, eventId);
+            } catch (AssertionError as) {
+                ResponseEntity.status(HttpStatus.BAD_REQUEST).body(as.getMessage());
+            }
+            return ResponseEntity.status(HttpStatus.OK).build();
         }
-        return ResponseEntity.status(HttpStatus.OK).build();
     }
 
     //Put, Get and Delete for favourites of a user
 
     @PutMapping("/users/{id}/favourites/{eventId}")
     public ResponseEntity<String> addFavourites(@PathVariable Long id, @PathVariable Long eventId) {
-        try {
-            favouriteService.addFavourite(id,eventId);
-        } catch (AssertionError as) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(as.getMessage());
+        if(!isCurrentUser(id)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        } else {
+            try {
+                favouriteService.addFavourite(id, eventId);
+            } catch (AssertionError as) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(as.getMessage());
+            }
+            userTrophyService.firstFavourite(id);
+            return ResponseEntity.status(HttpStatus.CREATED).body("Favourite added");
         }
-        return ResponseEntity.status(HttpStatus.CREATED).body("Favourite added");
     }
 
     @GetMapping("/users/{id}/favourites")
@@ -147,52 +176,70 @@ public class UserController {
 
     @DeleteMapping("/users/{id}/favourites/{eventId}")
     public ResponseEntity<String> removeFavourites(@PathVariable Long id, @PathVariable Long eventId) {
-        try {
-            favouriteService.removeFavourite(id, eventId);
-        } catch (AssertionError as) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(as.getMessage());
+        if(!isCurrentUser(id)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        } else {
+            try {
+                favouriteService.removeFavourite(id, eventId);
+            } catch (AssertionError as) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(as.getMessage());
+            }
+            return ResponseEntity.status(HttpStatus.OK).build();
         }
-        return ResponseEntity.status(HttpStatus.OK).build();
     }
 
     //Put, Get and Delete for trophies of a user
 
     @PutMapping("/users/{id}/trophies")
-    public ResponseEntity<String> addTrophies(@PathVariable Long id, @RequestBody List<Long> trophiesIds) {
-        try {
-            userTrophyService.addTrophy(id,trophiesIds);
-        } catch (AssertionError as) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(as.getMessage());
+    public ResponseEntity<UserDto> addTrophies(@PathVariable Long id, @RequestBody List<Long> trophiesIds) {
+        User user;
+        if(!isCurrentUser(id)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        } else {
+            try {
+                user = userTrophyService.addTrophy(id, trophiesIds);
+            } catch (AssertionError as) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+            }
+            return ResponseEntity.status(HttpStatus.CREATED).body(convertUserToDto(user));
         }
-        return ResponseEntity.status(HttpStatus.CREATED).body("Trophies added");
     }
 
     @GetMapping("/users/{id}/trophies")
     public ResponseEntity<List<TrophyDto>> getTrophies(@PathVariable Long id) {
-        List<Trophy> trophies = userTrophyService.getTrophiesByID(id);
-        return ResponseEntity.status(HttpStatus.OK).body(trophies.stream().map(this::convertTrophyToDto).toList());
+        List<Trophy> trophies = userTrophyService.getTrophiesById(id);
+        return ResponseEntity.status(HttpStatus.OK).body(trophies.stream().map(trophyMapper::convertTrophyToDto).toList());
     }
 
     @DeleteMapping("/users/{id}/trophies")
-    public ResponseEntity<String> removeTrophies(@PathVariable Long id, @RequestBody List<Long> trophiesIds) {
-        try {
-            userTrophyService.removeTrophy(id,trophiesIds);
-        } catch (AssertionError as) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(as.getMessage());
+    public ResponseEntity<UserDto> removeTrophies(@PathVariable Long id, @RequestBody List<Long> trophiesIds) {
+        User user;
+        if(!isCurrentUser(id)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        } else {
+            try {
+                user = userTrophyService.removeTrophy(id, trophiesIds);
+            } catch (AssertionError as) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+            }
+            return ResponseEntity.status(HttpStatus.OK).body(convertUserToDto(user));
         }
-        return ResponseEntity.status(HttpStatus.OK).build();
     }
 
     //Put, Get and Delete for the friend requests to other users of a user
 
     @PutMapping("/users/{id}/friends/{friendId}")
     public ResponseEntity<String> addRequestsTo(@PathVariable Long id, @PathVariable Long friendId) {
-        try {
-            requestService.addFriend(id,friendId);
-        } catch (AssertionError as) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(as.getMessage());
+        if(!isCurrentUser(id)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        } else {
+            try {
+                requestService.addFriend(id, friendId);
+            } catch (AssertionError as) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(as.getMessage());
+            }
+            return ResponseEntity.status(HttpStatus.CREATED).body("Request added");
         }
-        return ResponseEntity.status(HttpStatus.CREATED).body("Request added");
     }
 
     @GetMapping("users/{id}/friends")
@@ -206,13 +253,75 @@ public class UserController {
 
     @DeleteMapping("/users/{id}/friends/{friendId}")
     public ResponseEntity removeRequestsTo(@PathVariable Long id, @PathVariable Long friendId) {
-        try {
-            requestService.removeFriend(id,friendId);
-        } catch (AssertionError as) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(as.getMessage());
+        if(!isCurrentUser(id)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        } else {
+            try {
+                requestService.removeFriend(id, friendId);
+            } catch (AssertionError as) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(as.getMessage());
+            }
+            return ResponseEntity.status(HttpStatus.OK).build();
         }
-        return ResponseEntity.status(HttpStatus.OK).build();
+    }
 
+    @GetMapping("/users/generate_route")
+    public ResponseEntity<List<EventDto>> getRoutes(
+            @RequestParam(required = true) double lat,
+            @RequestParam(required = true) double lon,
+            @RequestParam(required = true) int radius,
+            @RequestParam(required = true) String day,
+            @RequestParam(required = false) Long userId
+    ) {
+        List<Event> events;
+        if(userId != null) {
+            events = routeService.getUserAdjustedRouteInADayAndLocation(lat, lon, radius, day, userId);
+            userTrophyService.firstRoute(userId);
+        } else {
+            events = routeService.getRouteInADayAndLocation(lat, lon, radius, day);
+        }
+        return ResponseEntity.status(HttpStatus.OK).body(events.stream().map(eventMapper::convertEventToDto).toList());
+    }
+
+    @GetMapping("/users/{id}/routes")
+    public ResponseEntity<List<RouteDto>> getRoutes(@PathVariable Long id) {
+        if(!isCurrentUser(id)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        } else {
+            List<Route> routes = routeService.getRoutes(id);
+            return ResponseEntity.status(HttpStatus.OK).body(routes.stream().map(routeMapper::convertRouteToDto).toList());
+        }
+    }
+
+    @PostMapping("/users/{id}/routes")
+    public ResponseEntity<RouteDto> saveRoute(@PathVariable Long id, @RequestBody List<EventDto> events) {
+        if(!isCurrentUser(id)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        } else {
+            List<Event> ev = events.stream().map(eventMapper::convertEventDtoToEntity).toList();
+            Route result = routeService.saveRoute(ev,id);
+            return ResponseEntity.status(HttpStatus.CREATED).body(routeMapper.convertRouteToDto(result));
+        }
+    }
+
+    @DeleteMapping("/users/{userId}/routes/{routeId}")
+    public ResponseEntity getEventsRoute(@PathVariable Long userId, @PathVariable Long routeId) {
+        if(!isCurrentUser(userId)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        } else {
+            try {
+                routeService.deleteRoute(routeId,userId);
+            } catch (AssertionError as) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(as.getMessage());
+            }
+            return ResponseEntity.status(HttpStatus.OK).build();
+        }
+    }
+
+    @GetMapping("/users/{userId}/reviews")
+    public ResponseEntity<List<ReviewDto>> getReviews(@PathVariable Long userId) {
+        List<Review> reviews = reviewService.getReviewsByUser(userId);
+        return ResponseEntity.status(HttpStatus.OK).body(reviews.stream().map(reviewMapper::convertReviewToDto).toList());
     }
 
     //Dto conversion functions
@@ -242,27 +351,18 @@ public class UserController {
         return userDto;
     }
 
-
     private User convertUserDtoToEntity(UserDto userDto) {
         if (userDto.getRole() == Role.ORGANIZER) {
             return  modelMapper.map(userDto, Organizer.class);
         }
         else return modelMapper.map(userDto, User.class);
-        //....modifications....
-
     }
 
-
-    private TrophyDto convertTrophyToDto(Trophy trophy) {
-        TrophyDto trophyDto = modelMapper.map(trophy, TrophyDto.class);
-        //....modifications....
-        return trophyDto;
-    }
-
-    private Trophy convertTrophyDtoToEntity(TrophyDto trophyDto) {
-        Trophy trophy = modelMapper.map(trophyDto, Trophy.class);
-        //....modifications....
-        return trophy;
+    //Utils
+    private boolean isCurrentUser(Long userId) {
+        CurrentUser currentUser = (CurrentUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        User u = userService.getUserByUsername(currentUser.getUsername());
+        return u.getId()==userId;
     }
 
 }
